v0.9.1 TimeTooSkewed more cases
PROBLEM:
Amazon SDK can't handle some cases with RequestClockTooSkewed because it uses current clock offset for calculating new offset too.

SOLUTION:
Reset offset when system clock changes

PROBLEM:
OBS3ExceptionHandler doesn't fix RequestClockTooSkewed problem in some cases

SOLUTION:
Do not relay on `- (void)URLSession:(NSURLSession *)session dataTask:(NSURLSessionDataTask *)task didReceiveData:(nonnull NSData *)receivedData`. It may not be called.

+++ Pod/Classes/AWS/OBS3ExceptionHandler.m
@@ -48,7 +48,7 @@ static NSString *MIMEApplicationXML = @"application/xml";
         {
             NSXMLParser *parser = [[NSXMLParser alloc] initWithData:data];
             S3ErrorResponseHandler *errorHandler =
-            [[S3ErrorResponseHandler alloc] initWithStatusCode:response.statusCode];
+            [[S3ErrorResponseHandler alloc] initWithStatusCode:(int32_t)response.statusCode];
             [parser setDelegate:errorHandler];
             [parser parse];
